Deleting views added

+++ Core/Datainterface/Delete.php
@@ -5,9 +5,12 @@ namespace Datainterface;
 class Delete
 {
     public static function delete($tableName, $keyValue = []): bool{
-        $helper = new CrudHelper();
-        $helper->setTableName($tableName);
-        return $helper->delete($keyValue);
+        if(SecurityChecker::isConfigExist()){
+            $helper = new CrudHelper();
+            $helper->setTableName($tableName);
+            return $helper->delete($keyValue);
+        }
+        return false;
     }
 
 }
